# src/docs/swagger.yaml
definitions:
...
  dto.ReqCreateUpdateTag:
    properties:
      about:
        type: string
      color:
        type: string
      id:
        type: integer
      name:
        type: string
    required:
    - name
    type: object
...
  dto.RespTag:
    properties:
      about:
        type: string
      color:
        type: string
      id:
        type: integer
      name:
        type: string
      user_id:
        type: integer
    type: object
...
  echo.HTTPError:
    properties:
      message: {}
    type: object
...
  pkg.Response:
    properties:
      body: {}
    type: object
info:
  contact: {}
paths:
...
  /tag/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an tag
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "403":
          description: invalid csrf
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: can't find tag with such id
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: invalid http method
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Delete an tag
      tags:
      - tag
    get:
      consumes:
      - application/json
      description: Get tag by id
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success get tag
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                body:
                  $ref: '#/definitions/dto.RespTag'
              type: object
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: invalid http method
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Show a post
      tags:
      - tag
  /tag/create:
    post:
      consumes:
      - application/json
      description: Create tag
      parameters:
      - description: tag info
        in: body
        name: tag
        required: true
        schema:
          $ref: '#/definitions/dto.ReqCreateUpdateTag'
      produces:
      - application/json
      responses:
        "200":
          description: success update tag
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                body:
                  $ref: '#/definitions/dto.RespTag'
              type: object
        "400":
          description: bad req
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "403":
          description: invalid csrf or permission denied
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: invalid http method
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "422":
          description: unprocessable entity
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Create tag
      tags:
      - tag
  /tag/edit:
    post:
      consumes:
      - application/json
      description: Update an tag
      parameters:
      - description: tag info
        in: body
        name: tag
        required: true
        schema:
          $ref: '#/definitions/dto.ReqCreateUpdateTag'
      produces:
      - application/json
      responses:
        "200":
          description: success update tag
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                body:
                  $ref: '#/definitions/dto.RespTag'
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "403":
          description: invalid csrf or permission denied
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: invalid http method
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "422":
          description: unprocessable entity
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Update an tag
      tags:
      - tag
...
swagger: "2.0"